+++ Server/src/main/java/JGS/CasperEvent/domain/event/dto/request/lotteryEventDto/CasperBotRequestDto.java
@@ -1,4 +1,4 @@
-package JGS.CasperEvent.domain.event.dto.RequestDto.lotteryEventDto;
+package JGS.CasperEvent.domain.event.dto.request.lotteryEventDto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.Max;
